packages/python/stoopid-logging/pyproject.toml
[project]
name = "stoopid-logging"
version = "0.0.0"
description = "Structured JSON logging for Python services, conforming to the stoopid-commons shared log-event schema."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Stoopid Company" }]
keywords = ["logging", "structlog", "json", "opentelemetry", "structured"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Typing :: Typed",
]
dependencies = [
    "structlog>=24.4,<26",
]

[project.optional-dependencies]
otel = ["opentelemetry-api>=1.25,<2"]

[project.urls]
Homepage = "https://github.com/stoopidJSON/stoopid-commons"
Source = "https://github.com/stoopidJSON/stoopid-commons/tree/main/packages/python/stoopid-logging"
Issues = "https://github.com/stoopidJSON/stoopid-commons/issues"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stoopid_logging"]

[tool.hatch.build.targets.sdist]
include = ["src/stoopid_logging", "tests", "README.md", "CHANGELOG.md", "LICENSE"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "jsonschema>=4.21",
    "opentelemetry-api>=1.25,<2",
    "opentelemetry-sdk>=1.25,<2",
]
